requirements.txt
fastapi
uvicorn[standard]
python-multipart
passlib
opencv-python
numpy
Pillow
huggingface_hub
python-dotenv
PyJWT
